+++ spec/javascripts/components/story/collapsed_story/collapsed_story_estimate_button_spec.js
@@ -4,31 +4,31 @@ import { CollapsedStoryEstimateButton } from 'components/story/CollapsedStory/Co
 
 describe('<CollapsedStoryEstimateButton />', () => {
   it("renders component with 'Fibonacci' point scale", () => {
-    const props = { pointValues: ['1','2','3','5','8'] }
-    const wrapper = shallow(<CollapsedStoryEstimateButton project={props}/>);
+    const project = { pointValues: ['1','2','3','5','8'] }
+    const wrapper = shallow(<CollapsedStoryEstimateButton project={project}/>);
     const text = wrapper.text()
 
-    props.pointValues.forEach((value)=>{
+    project.pointValues.forEach((value)=>{
       expect(text).toContain(value)
     });
   });
 
   it("renders component with 'Powers of two' point scale", () => {
-    const props = { pointValues: ['1','2','4','8'] }
-    const wrapper = shallow(<CollapsedStoryEstimateButton project={props}/>);
+    const project = { pointValues: ['1','2','4','8'] }
+    const wrapper = shallow(<CollapsedStoryEstimateButton project={project}/>);
     const text = wrapper.text()
 
-    props.pointValues.forEach((value)=>{
+    project.pointValues.forEach((value)=>{
       expect(text).toContain(value)
     });
   });
 
   it("renders component with 'Linear' point scale", () => {
-    const props = { pointValues: ['1','2','3','4','5'] }
-    const wrapper = shallow(<CollapsedStoryEstimateButton project={props}/>);
+    const project = { pointValues: ['1','2','3','4','5'] }
+    const wrapper = shallow(<CollapsedStoryEstimateButton project={project}/>);
     const text = wrapper.text()
 
-    props.pointValues.forEach((value)=>{
+    project.pointValues.forEach((value)=>{
       expect(text).toContain(value)
     });
   });