+++ CHANGELOG.md
@@ -16,7 +16,7 @@
 - Added Popover when Story is collapsed in board V2
 
 ### Changed
-- Change estimate button to dinamic values when story is collapsed in board V2
+- Change estimate button to dynamic values when story is collapsed in board V2
 
 ## [1.21.1] 2018-10-18
 ### Fixed